reports-service/requirements.txt
fastapi
uvicorn[standard]
clickhouse-driver
requests
PyJWT==2.8.0